genericized authorization middleware
