+++ server/app/routes/index.js
@@ -15,3 +15,6 @@ router.use('/reviews', require('./reviews'));
 router.use(function (req, res) {
     res.status(404).end();
 });
+
+
+
